@@ -5,25 +5,24 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": "^7.2.5",
-        "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "^1.0",
-        "guzzlehttp/guzzle": "^6.3",
-        "laravel/framework": "^7.0",
-        "laravel/socialite": "^4.3",
-        "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
-        "tymon/jwt-auth": "^1.0.0"
+        "php": "^7.3|^8.0",
+        "fideloper/proxy": "^4.4",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^8.12",
+        "laravel/socialite": "^5.0",
+        "laravel/tinker": "^2.5",
+        "laravel/ui": "^3.0",
+        "tymon/jwt-auth": "^1.0.1"
     },
     "require-dev": {
-        "beyondcode/laravel-dump-server": "^1.4.0",
-        "doctrine/dbal": "^2.9",
-        "facade/ignition": "^2.0",
-        "fzaninotto/faker": "^1.9.1",
-        "laravel/dusk": "^5.11",
-        "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "doctrine/dbal": "^2.12",
+        "facade/ignition": "^2.5",
+        "fakerphp/faker": "^1.9.1",
+        "laravel/dusk": "^6.8",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.3.3"
     },
     "config": {
         "optimize-autoloader": true,
@@ -39,12 +38,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+          "App\\": "app/",
+          "Database\\Factories\\": "database/factories/",
+          "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
