@@ -2,7 +2,7 @@
 
 namespace Tests\Browser;
 
-use App\User;
+use App\Models\User;
 use Tests\Browser\Pages\Home;
 use Tests\Browser\Pages\Login;
 use Tests\DuskTestCase;
@@ -19,7 +19,7 @@ class LoginTest extends DuskTestCase
     /** @test */
     public function login_with_valid_credentials()
     {
-        $user = factory(User::class)->create();
+        $user = User::factory()->make();
 
         $this->browse(function ($browser) use ($user) {
             $browser->visit(new Login)
@@ -41,7 +41,7 @@ class LoginTest extends DuskTestCase
     /** @test */
     public function log_out_the_user()
     {
-        $user = factory(User::class)->create();
+        $user = User::factory()->make();
 
         $this->browse(function ($browser) use ($user) {
             $browser->visit(new Login)
