@@ -10,12 +10,12 @@ class Authenticate extends Middleware
      * Get the path the user should be redirected to when they are not authenticated.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return string
+     * @return string|null
      */
     protected function redirectTo($request)
     {
         if (! $request->expectsJson()) {
-            return redirect('login');
+            return redirect('/login');
         }
     }
 }
